Rename reducer for brevity
The specific purpose of the reducer doesn't need to be stated. It can be
kept generic if we want to set any given columnbounds on a columnbounds
input.
