+++ packages/harpguru-core/src/components/callback-on-sourceglobalprops/utils/reduce-columnbounds/index.ts
@@ -1,0 +1,1 @@
+export { reduceColumnBounds } from './reduce-columnbounds'
