--- packages/harpguru-core/src/components/callback-on-sourceglobalprops/utils/reduce-sourcecolumnbounds-to-columnbounds/index.ts
@@ -1,1 +1,0 @@
-export { reduceSourceColumnBoundsToColumnBounds } from './reduce-sourcecolumnbounds-to-columnbounds'
